c# GUI 수정
c# GUI 수정

+++ XYSTAGE_SYSTEM_C#/XYSTAGE_SYSTEM_DESIGN/CTwincat.cs
@@ -139,7 +139,7 @@ namespace XYSTAGE_SYSTEM_DESIGN
         #endregion
 
         #region Set Position, Velocity , Acceleration , Deceleration
-        public void SetPos(int pos)
+        public void SetPos(double pos)
         {
             if (IsConnect)
             {
